@@ -1,41 +1,38 @@
-﻿using System.Collections.Generic;
-using System.Linq;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using RssFeeder.Mvc.Models;
 
-namespace RssFeeder.Mvc.Controllers
+namespace RssFeeder.Mvc.Controllers;
+
+[Authorize]
+public class FeedController : Controller
 {
-    [Authorize]
-    public class FeedController : Controller
+    private readonly string _sourceFile = "feeds.json";
+    private readonly List<FeedModel> _feeds;
+
+    public FeedController()
     {
-        private readonly string _sourceFile = "feeds.json";
-        private readonly List<FeedModel> _feeds;
+        _feeds = System.Text.Json.JsonSerializer.Deserialize<List<FeedModel>>(
+            System.IO.File.ReadAllText(_sourceFile),
+            new System.Text.Json.JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new List<FeedModel>();
+    }
 
-        public FeedController()
-        {
-            _feeds = System.Text.Json.JsonSerializer.Deserialize<List<FeedModel>>(
-                System.IO.File.ReadAllText(_sourceFile),
-                new System.Text.Json.JsonSerializerOptions { PropertyNameCaseInsensitive = true });
-        }
+    // GET: Feed
+    public ActionResult Index()
+    {
+        return View(_feeds.OrderBy(i => i.title).AsEnumerable());
+    }
 
-        // GET: Feed
-        public ActionResult Index()
-        {
-            return View(_feeds.OrderBy(i => i.title).AsEnumerable());
-        }
+    [AllowAnonymous]
+    public IActionResult List()
+    {
+        return Json(_feeds);
+    }
 
-        [AllowAnonymous]
-        public IActionResult List()
-        {
-            return Json(_feeds);
-        }
-
-        // GET: Feed/Details/5
-        public ActionResult Details(string id)
-        {
-            var item = _feeds.FirstOrDefault(q => q.id == id);
-            return View(item);
-        }
+    // GET: Feed/Details/5
+    public ActionResult Details(string id)
+    {
+        var item = _feeds.FirstOrDefault(q => q.id == id);
+        return View(item);
     }
 }
