@@ -1,73 +1,72 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Web;
+﻿using System.Web;
 using Newtonsoft.Json;
 
-namespace RssFeeder.Mvc.Models
+namespace RssFeeder.Mvc.Models;
+
+public record ExportFeedItems : RssFeedItem
 {
-    public class ExportFeedItems : RssFeedItem
+    public ExportFeedItems() : base()
+    { }
+}
+
+public record RssFeedItem
+{
+    public RssFeedItem()
     {
-        public ExportFeedItems() : base()
-        { }
+        FeedAttributes = new FeedAttributes();
+        OpenGraphAttributes = new Dictionary<string, string>();
+        HtmlAttributes = new Dictionary<string, string>();
     }
 
-    public class RssFeedItem
+    [JsonProperty(PropertyName = "id")]
+    public string? Id { get; set; }
+    public string? FeedId { get; set; }
+    public string? Url { get; set; }
+    public string? UrlHash { get; set; }
+    public string? ImageUrl { get; set; }
+    public string? VideoUrl { get; set; }
+    public int VideoHeight { get; set; }
+    public int VideoWidth { get; set; }
+    public string? Title { get; set; }
+    public string? Subtitle { get; set; }
+    public string? ArticleText { get; set; }
+    public string? Description { get; set; }
+    public string? MetaDescription { get; set; }
+    public DateTime DateAdded { get; set; }
+    public string? FileName { get; set; }
+    public string? SiteName { get; set; }
+    public string? HostName { get; set; }
+    public string? LinkLocation { get; set; }
+    public Dictionary<string, string> OpenGraphAttributes { get; set; }
+    public Dictionary<string, string> HtmlAttributes { get; set; }
+    public FeedAttributes FeedAttributes { get; set; }
+
+    [JsonIgnore]
+    public string? EncodedDescription
     {
-        public RssFeedItem()
+        get
         {
-            FeedAttributes = new FeedAttributes();
-        }
-
-        [JsonProperty(PropertyName = "id")]
-        public string Id { get; set; }
-        public string FeedId { get; set; }
-        public string Url { get; set; }
-        public string UrlHash { get; set; }
-        public string ImageUrl { get; set; }
-        public string VideoUrl { get; set; }
-        public int VideoHeight { get; set; }
-        public int VideoWidth { get; set; }
-        public string Title { get; set; }
-        public string Subtitle { get; set; }
-        public string ArticleText { get; set; }
-        public string Description { get; set; }
-        public string MetaDescription { get; set; }
-        public DateTime DateAdded { get; set; }
-        public string FileName { get; set; }
-        public string SiteName { get; set; }
-        public string HostName { get; set; }
-        public string LinkLocation { get; set; }
-        public Dictionary<string, string> OpenGraphAttributes { get; set; }
-        public Dictionary<string, string> HtmlAttributes { get; set; }
-        public FeedAttributes FeedAttributes { get; set; }
-
-        [JsonIgnore]
-        public string EncodedDescription
-        {
-            get
-            {
-                return HttpUtility.HtmlEncode(Description);
-            }
-        }
-
-        [JsonIgnore]
-        public string FormattedDateAdded
-        {
-            get
-            {
-                return DateAdded.ToString("ddd, dd MMM yyyy HH':'mm':'ss 'GMT'");
-            }
+            return HttpUtility.HtmlEncode(Description);
         }
     }
 
-    public class FeedAttributes
+    [JsonIgnore]
+    public string FormattedDateAdded
     {
-        public string FeedId { get; set; }
-        public string Url { get; set; }
-        public string UrlHash { get; set; }
-        public string Title { get; set; }
-        public DateTime DateAdded { get; set; }
-        public string FileName { get; set; }
-        public string LinkLocation { get; set; }
+        get
+        {
+            return DateAdded.ToString("ddd, dd MMM yyyy HH':'mm':'ss 'GMT'");
+        }
     }
+}
+
+public record FeedAttributes
+{
+    public string? FeedId { get; set; }
+    public string? Url { get; set; }
+    public string? UrlHash { get; set; }
+    public string? Title { get; set; }
+    public DateTime DateAdded { get; set; }
+    public string? FileName { get; set; }
+    public string? LinkLocation { get; set; }
 }
