@@ -1,16 +1,13 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using Microsoft.Azure.Cosmos;
+﻿using Microsoft.Azure.Cosmos;
 using RssFeeder.Mvc.Models;
 
-namespace RssFeeder.Mvc.Services
+namespace RssFeeder.Mvc.Services;
+
+public interface IDatabaseService
 {
-    public interface IDatabaseService
-    {
-        Task<IEnumerable<RssFeedItem>> GetItemsAsync(QueryDefinition queryDef);
-        Task<RssFeedItem> GetItemAsync(string id);
-        Task AddItemAsync(RssFeedItem item);
-        Task UpdateItemAsync(string id, RssFeedItem item);
-        Task DeleteItemAsync(string id);
-    }
+    Task<IEnumerable<RssFeedItem>> GetItemsAsync(QueryDefinition queryDef);
+    Task<RssFeedItem> GetItemAsync(string id);
+    Task AddItemAsync(RssFeedItem item);
+    Task UpdateItemAsync(string id, RssFeedItem item);
+    Task DeleteItemAsync(string id);
 }
