@@ -17,9 +17,8 @@ public class TestCommand : OaktonCommand<TestInput>
         var webUtils = _container.Resolve<IWebUtils>();
         var parser = _container.ResolveNamed<ITagParser>("htmltag-parser");
 
-        string url = "https://media.townhall.com/townhall/reu/o/2021/224/171327fc-9d62-429d-b6c3-ede53197355a-1110x740.jpg";
-        //string url = "https://pjmedia.com/columns/ari-j-kaufman/2021/08/12/what-did-san-francisco-expect-when-it-elected-the-progeny-of-of-militant-marxist-terrorists-n1468973";
-        var response = webUtils.TrySaveUrlToDisk(url, "hash", "test.jpg");
+        string url = "https://nationalpost.com/pmn/entertainment-pmn/king-charles-vows-to-serve-his-nation-as-britain-mourns-late-queen";
+        var response = webUtils.WebDriverUrlToDisk(url, "test.html");
 
         Log.Information("Response = {response}", response);
         Log.CloseAndFlush();
