@@ -1,17 +1,11 @@
-﻿using System;
-using System.IO;
-using System.Linq;
-using System.Reflection;
-using Microsoft.Extensions.Hosting;
-
 public class AppVersionInfo
 {
     private static readonly string _buildFileName = ".buildinfo.json";
-    private string _buildFilePath;
-    private string _buildNumber;
-    private string _gitHash;
-    private string _gitShortHash;
-    private string _workflow;
+    private string? _buildFilePath;
+    private string? _buildNumber;
+    private string? _gitHash;
+    private string? _gitShortHash;
+    private string? _workflow;
 
     public AppVersionInfo(IHostEnvironment hostEnvironment)
     {
@@ -80,7 +74,7 @@ public class AppVersionInfo
                 }
             }
 
-            return _gitHash;
+            return _gitHash!;
         }
     }
 
@@ -110,7 +104,7 @@ public class AppVersionInfo
                 }
             }
 
-            return _workflow;
+            return _workflow!;
         }
     }
 
