@@ -1,12 +1,9 @@
-﻿using System;
+namespace RssFeeder.Mvc.Models;
 
-namespace RssFeeder.Mvc.Models
+public record Agent
 {
-    public class Agent
-    {
-        public string BrowserAgent { get; set; }
-        public string IpAddress { get; set; }
-        public string Referrer { get; set; }
-        public DateTime Timestamp { get; set; }
-    }
+    public string? BrowserAgent { get; set; }
+    public string? IpAddress { get; set; }
+    public string? Referrer { get; set; }
+    public DateTime Timestamp { get; set; }
 }
