@@ -1,4 +1,4 @@
-﻿@using System.Security.Principal
+@using System.Security.Principal
 
 <ul class="navbar-nav">
     @if (User.Identity.IsAuthenticated)
